#514 Added confirmation while deleting component on a grahpical view

+++ WebContent/WEB-INF/jsp/viewEdit.jsp
@@ -195,52 +195,55 @@
         graphicRendererEditor.close();
         staticEditor.close();
         compoundEditor.close();
+        customEditor.close();
     }
-    
+
     function updateViewComponentLocation(divId) {
         var div = $(divId);
         var lt = div.style.left;
         var tp = div.style.top;
-        
+
         // Remove the 'px's from the positions.
         lt = lt.substring(0, lt.length-2);
         tp = tp.substring(0, tp.length-2);
-        
+
         // Save the new location.
         ViewDwr.setViewComponentLocation(div.viewComponentId, lt, tp);
     }
-    
+
     function addDnD(divId) {
         var div = $(divId);
         var dragSource = new dojo.dnd.HtmlDragMoveSource(div);
         dragSource.constrainTo($("viewBackground"));
-        
+
         // Save the drag source in the div in case it gets deleted. See below.
         div.dragSource = dragSource;
         // Also, create a function to call on drag end to update the point view's location.
         div.onDragEnd = function() {updateViewComponentLocation(divId);};
-        
+
         dojo.event.connect(dragSource, "onDragEnd", div.onDragEnd);
     }
-    
+
     function deleteViewComponent(viewComponentId) {
         closeEditors();
-        ViewDwr.deleteViewComponent(viewComponentId);
-        
-        var div = $("c"+ viewComponentId);
-        
-        // Unregister the drag source from the DnD manager.
-        div.dragSource.unregister();
-        // Disconnect the event handling for drag ends on this guy.
-        $("viewContent").removeChild(div);
+        if(confirm('<fmt:message key="common.confirmDelete"/>')) {
+            ViewDwr.deleteViewComponent(viewComponentId);
+
+            var div = $("c"+ viewComponentId);
+
+            // Unregister the drag source from the DnD manager.
+            div.dragSource.unregister();
+            // Disconnect the event handling for drag ends on this guy.
+            $("viewContent").removeChild(div);
+        }
     }
-    
+
     function getViewComponentId(node) {
         while (!(node.viewComponentId))
             node = node.parentNode;
         return node.viewComponentId;
     }
-    
+
     function iconizeClicked() {
         ViewDwr.getViewComponentIds(function(ids) {
             var i, comp, content;
@@ -260,7 +263,7 @@
                     comp = $("c"+ ids[i]);
                     content = $("c"+ ids[i] +"Content");
                     if (comp.savedState)
-                        mango.view.setContent(comp.savedState);                
+                        mango.view.setContent(comp.savedState);
                     else if (comp.savedContent)
                         content.innerHTML = comp.savedContent;
                     else
@@ -278,12 +281,12 @@
 
 		if(width > currentWidth) {
 			$("viewBackground").width = parseInt(width,10) + 30;
-		} 
+		}
 		if(height > currentHeight) {
 			$("viewBackground").height = parseInt(height,10) + 30;
 		}
 	}
-	
+
 	function resizeViewBackgroundToResolution(size) {
 		if(document.getElementById("viewBackground").src.includes("spacer.gif")){
 			switch(size) {
@@ -313,11 +316,11 @@
 			}
         } else {
         	document.getElementById("view.resolution").style.visibility = 'hidden';
-        	document.getElementById("sizeLabel").style.visibility = 'hidden';        	
+        	document.getElementById("sizeLabel").style.visibility = 'hidden';
         }
-		
+
 	}
-	
+
 	function deleteConfirm(){
 		if(document.getElementById("deleteCheckbox").checked) {
 			document.getElementById("deleteButton").style.visibility = 'visible';
@@ -326,7 +329,7 @@
 				document.getElementById("deleteButton").style.visibility = 'hidden';
 			}, 3000);
 		} else {
-			document.getElementById("deleteButton").style.visibility = 'hidden'; 
+			document.getElementById("deleteButton").style.visibility = 'hidden';
 		}
 	}
